move autoscale into separate module still called by scaleset code
